refacto

+++ src/main/java/com/kobe/warehouse/service/tiketz/dto/TicketZCreditProjection.java
@@ -1,3 +1,3 @@
 package com.kobe.warehouse.service.tiketz.dto;
 
-public record TicketZCreditProjection(Long userId, String firstName, String lastName, Long montant) {}
+public record TicketZCreditProjection(Integer userId, String firstName, String lastName, Long montant) {}
